avances varios

+++ reservas-core/src/main/java/com/reservas/dao/PerfilDAO.java
@@ -8,6 +8,6 @@ import com.reservas.model.PerfilBO;
  * Fecha: 2017-07-22 
  * Copyright 2017
  */
-public interface PerfilDAO extends BaseDAO<Long, PerfilBO>{
+public interface PerfilDAO extends BaseDAO<Integer, PerfilBO>{
 
 }
